Start ExternalData infrastructure.  Add logo.
Logo created with Gimp.  The exported logo could put into the ExternalData
mechanism because the configuration is relative to the conf.py file, which lives
in the source tree.

ExternalData CMake the same as ITK, but with the updated feature that you can
specify where the downloaded content is place.  We specify the root of the build
directory so that all the content links live in their relative location in the
source tree.