Add script for creating tarballs.
Create a tar.gz and .zip for each example.

We are in a bit of a pickles because sphinx needs the images to be present with
the .rst, and the example test code when exported needs to be able to find the
input images.  Putting the images in the source tree is a poor option.  For now,
at least, we put the input images in a "build" directory of the tarball.  This
way, when the example is built in the build directory, the images will be
available.

Also, I have to renege on build dirhtml instead of html -- dirhtml creates an
extra, unneeded directory.