Fix tarball input image path specification.
We replace CMAKE_CURRENT_BINARY_DIR with CMAKE_CURRENT_SOURCE_DIR when we
generate the tarballs.  This way we do have to put the input images in the build
directory of the tarball, and the user can nuke the contents of their build
directory as usual, and tests will past the next time around.