Start documentation on how to write an example.
And fix the example, and the rest of the repository to follow along.

Also fix pdflatex building.  The file globbing was not finding the file on
initial configuration because it did not exist yet.  Move to main CMakeLists.txt
with file specification because other PDF generation mechanisms than pdflatex
may be used with the latex output.