ENH: Remove dropbox account references. #9.
Remove Dropbox references for the Doxygen files and the ExternalData
resource.  The ExternalData resource is no longer needed because everything
is on the Midas server.  The Doxygen HTML and tag resources are replaced
by the nightlies now on Kitware's servers.  A Python gunzip is created to
unzip the tag file.  The doxylink Sphinx extension was updated to the latest
version because it was possibly having issues with the new version of the
Doxygen tag file.