ENH: Add infrastructure for Python tests
The baseline comparison is done automatically if there is a Python
test which was run.

The ITK_WRAP_PYTHON variable is fetched through the ITKConfig.cmake file,
there is no performance impact by doing this as find_package(ITK) is called
anyway for each example.

Older versions of ITK do not have the ITK_WRAP_PYTHON variable in their
config file: in this case the Python tests will not be run.

The DilateABinaryImage and ErodeABinrayImage examples are used to show the new
example setup.

Coauthored with
matt.mccormick@kitware.com
arnaudgelas@gmail.com

Change-Id: I545c40c1d43adf6ab69cf5aca514353438121d18