BUG: error when TEST_IMAGE_PREFIX was used
At CMake time, it was looking for ${TEST_IMAGE_PREFIX}.* files in the example
binary directory, and the corresponding file was used for the baseline
comparison test. Files were not found in an empty binary tree.

The solution here is to provide the full name for the the test image. Thus
instead of using TEST_IMAGE_PREFIX in compare_to_baseline, one should rather
use TEST_IMAGE.

Change-Id: I5cb15314c9871e4d3aab966f99232f3a8b40dfcc