COMP: Fix missing images in ApplyGradientRecursiveGaussianFilter
The images need to be the ones moved to the _images folder to be
displayed online. The example would correctly display the images locally,
but on the cdash setup the OutputX.png, OutputY.png and OutputMagnitude.png
images are not copied, so we need to use the baseline images instead.

Change-Id: I799f58d256e95d193e3f2ec26ffde40675cd7bb8