COMP: Set CMP0054 to NEW.
Also refactor the other cmake_policy calls.

This addresses the warning:

CMake Warning (dev) at CMakeLists.txt:11 (if):
Policy CMP0054 is not set: Only interpret if() arguments as variables or
keywords when unquoted.  Run "cmake --help-policy CMP0054" for policy
details.  Use the cmake_policy command to set the policy and suppress
this warning.

Quoted variables like "MSVC" will no longer be dereferenced when the
policy is set to NEW.  Since the policy is not set the OLD behavior will
be used. This warning is for project developers.  Use -Wno-dev to suppress
it.

Change-Id: I687af4d219fd75e9efaae33b52852bb4d5625a45