ENH: Add Python superbuild.
This add CPython to the superbuild.  It uses the python-cmake-buildsystem
repository, and the ExternalProject configuration is based off the Slicer
configuration, Slicer commit 3eb19c2.

Python requires zlib, so the cmake-ified CommonTK name mangled version is
also built. This is also passed to other projects.

ITK is built with static libraries, which addresses runtime library location
of the wrapping on Windows.

Change-Id: I97d44180d2b1b88e1ee76772a1407fa8f5608242