ENH: Move the Download button location.
The Download button location is moved to the right of the Previous and Next
example buttons.  Since the Download button is only available for examples,
and it is not available in the Module and Group overview pages, the Previous
and Next buttons move locations when hitting these pages. This makes it more
difficult to click through examples by continuously clicking the Previous or
Next buttons. Moving the Download button to the right keeps the Previous and
Next buttons in the same location.

Change-Id: Iaa533ede81fbb05af8bfc87d6b1c80a2254c83b1