ENH: Add ConvertAnitkImageTovtkImageData example.
We do not add a test for the Python version because the superbuild uses a VTK
build tree and it is currently difficult to use vtk's Python wrapping with a build
tree VTK that has other Python packages like itk.  The Python script was
tested with install tree vtk Python wrapping.

Change-Id: I6a7327119e17256cd6d2fc2783e6d1cdbeca9d15