BUG: Fix LineSpatialObject example
This example should be renamed to LineSpatialObject, not LineSpatialObject2.

In addition, the breathelink has been suppressed for a future commit.
Once the documentation is pushed to the official itk.org, this can be
uncommnented.

The reason for this is because the doxygen tries to find a
url that is not present. Hence, once it is pressent the error will be
fixed.