BUG: Incorrect bounds for python scripts in documentation
Many of the examples that contain python examples, have incorrect bounds of the lines included in the examples. Essentially, it is removing needed imports for the python example or it contains comments from the NUMFocus (which do not need to be included in the documentation).

In addition, the Documentation.rst template has been updated to remove only the NUMFocus.