ENH: Allow examples to be built without QuickView
Presently, all of the examples that utilize QuickView are disabled if
QuickView is not enabled. However, these examples strictly use QuickView
for visualization purposes and don't contribute to the core
functionality of the program.

These changes primarily change 2 types of CMakeList.txt files: the ones
for a module and those for a specific example.

For modules, it will remove if(ENABLE_QUICKVIEW) to simply just enable
the example. If some of the examples strictly used VTK and not
QuickView, I replaced it with if(VTK_DIR).

For examples, it will first check if QuickView is enabled. If so, that
will find the VTK package and only utilize the components we need which
primarily is: vtkRenderingCore and vtkRenderingGL2PSOpenGL2. If the VTK
version is less than 8.9, then the VTK_USE_FILE is included. Finally, it
links the VTK_LIBRARIES, and intiates the VTK module if the VTK version is atleast 8.9.