ENH: Configure examples that utilize VTK
Many of the examples that utilize VTK were not configured correctly or
needed to be updated to accomodate VTK 9.1.0.

This primarily included, finding the VTK package and only using the
components that specific example needed. If the VTK version is less than
8.9, then the VTK_USE_FILE is included. Finally, the VTK_LIBRARIES are
linked and if the VTK version is atleast 8.9, then the module is
inialized using VTK's vtk_module_autoinit().