BUG: Fix input for examples that use VTK
After fixing the CMakeList.txt scipts for examples that use VTK library
in some form, many of the examples needed there test input and
documentation updated.

This primarily added an input image for an examples test, and updated
it's respective documentation.txt with the input and output.

A few of the examples and/or their respective tests contained errors
that were marked with a TODO, for a future PR. These included:
  1) Baseline comparisons, that cannot work due to the output not ever
     being stored. In a future PR, the examples should be modified to
     save their output, so we can test against it.
       + VTKImageToITKImageBaseline
       + InPlaceFilterOfImageBaseline
       + GeometricPropertiesOfRegionBaseline
  2) The DemonstrateThresholdAlgorithm example has it's
     KittlerIllingworth filter commented out with a TODO. This was the
     only filter to not work, and will be documented in a future issue.