DOC: Rename folder directories
Many of the example's folder names differ from the headers in their
respective Documentation.rst. In order to help facilitate the alphabetical ordering of these examples on Sphinx, this will make the headers (which are shown on Sphinx) nearly identical to their folder name.

Thus, when we look to add new examples in an index.rst, we can order
them alphebetically by folder name which will help the save time for the
developer.