COMP: Remove "vtk" prefix from VTK COMPONENTS
In modern VTK (> 8.90) the COMPONENT names don't have the "vtk" prefix.

This fix get the VTK version first, and then look for the COMPONENTS
with the right name, prepending ${_vtk_prefix} to the names.

The examples using ITKVtkGlue (`ENABLE_QUICKVIEW` = ON) re-use the variable
`_vtk_prefix` defined in the top CMakeLists.

The examples that depend on `VTK` and `ITK without ITKVTKGlue` set
`_vtk_prefix` by their own.

Fix #241